refactor(observability): relocate audit-cache writer out of materialise.py; fix stale barrier prose
- sink_audit / prune_audit_cache move verbatim to
  rwa_calc.observability.audit_cache (the 'sink_parquet only in
  materialise.py' invariant they were co-located for was never
  implemented in arch_check — verified phantom rule). 8 import sites
  updated; spill-file registry stays in materialise.py; behaviour
  identical.
- materialise.py docstring now states the verified plan-depth mechanism
  (recursive plan-tree depth, ~25,000-node measured threshold on Polars
  1.37, barriers also bound plan-construction time ~100x) instead of the
  stale '>500-node' folklore; points at
  docs/plans/single-lazy-plan-refactor.md.
- pipeline-collect-barriers.md: cpu (not streaming) documented as the
  default collect engine; every barrier line reference re-verified; the
  undocumented classifier_output barrier added to table + diagram.

Phase 0 of docs/plans/target-architecture-migration.md.
